requirements.txt
absl-py==2.0.0
altair==5.5.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
biopython==1.83
blinker==1.9.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.2.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.11.0
idna==3.4
imageio==2.35.1
intel-openmp==2021.4.0
ipykernel==6.29.2
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.9
jupyter_core==5.7.1
keras==3.5.0
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
narwhals==1.19.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.24.3
psutil==5.9.8
pyarrow==18.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.16.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spyder-kernels==2.3.3
streamlit==1.41.1
sympy==1.13.1
tbb==2021.13.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.1
tensorflow==2.17.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.5
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==2.3.7
wrapt==1.14.1
wsproto==1.2.0
